🔥 chore(.gitignore): add files and directories to ignore list Added files and directories to the ignore list to prevent them from being committed to the repository. This includes node_modules, .vscode, package-lock.json, old, json, and .env files.
🎉 feat(index.js): add script to slice and enrich data with db-ip API
This commit adds a script that reads data from JSON files, enriches it with data from the db-ip API, and slices it into smaller files. The script reads data from six different files, and for each file, it checks if the file exists before proceeding. If the file exists, it reads the data, enriches it with data from the db-ip API, and slices it into smaller files. The script writes the enriched data to new JSON files with the naming convention of {edge}-{start}-{end}-dbip.json.

🎉 feat(index.v3.js): add script to convert CSV to JSON and enrich data with dbip API
This commit adds a script that reads a CSV file, converts it to JSON, and enriches each record with data from the dbip API. The enriched data is then saved to multiple output files, each containing a maximum of 1000 records. The script uses the dotenv package to read the API key from the environment variables. A progress bar is displayed during the execution of the script to indicate the progress of the processing.

🆕 chore(package.json): add start script and node engine version
The start script has been added to the package.json file to allow the application to be started with the command 'npm start'. The node engine version has also been specified to version 19.